add model statistics modul

+++ microservice_2/access-control-service/src/main/java/com/example/accesscontrolservice/util/DataGenerator.java
@@ -14,6 +14,7 @@ import com.github.javafaker.Faker;
 import lombok.RequiredArgsConstructor;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import jakarta.annotation.PostConstruct;
 
@@ -43,9 +44,17 @@ public class DataGenerator {
     private final ObjectMapper objectMapper;
 
     private static final String reportPath = "../fingerprint_training/reports/";
+    
+    @Value("${data.generator.enabled:false}")
+    private boolean dataGeneratorEnabled;
+
     @PostConstruct
     public void initializeData() throws IOException {
-
+        if (!dataGeneratorEnabled) {
+            System.out.println("Data generator is disabled. Skipping data initialization.");
+            return;
+        }
+        
         accessLogRepository.deleteAll();
         cameraRepository.deleteAll();
         areaRepository.deleteAll();
