+++ microservice_2/user-management-service/src/main/java/com/example/usermanagementservice/util/DataGenerator.java
@@ -1,6 +1,6 @@
 package com.example.usermanagementservice.util;
 
-import com.example.usermanagementservice.model.Admin;;
+import com.example.usermanagementservice.model.Admin;
 import com.example.usermanagementservice.model.Employee;
 import com.example.usermanagementservice.repository.AdminRepository;
 import com.example.usermanagementservice.repository.EmployeeRepository;
@@ -11,6 +11,8 @@ import lombok.RequiredArgsConstructor;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.stereotype.Service;
+import org.springframework.beans.factory.annotation.Value;
+
 import jakarta.annotation.PostConstruct;
 
 import java.io.File;
@@ -39,8 +41,17 @@ public class DataGenerator {
     private final ObjectMapper objectMapper;
 
     private static final String reportPath = "../fingerprint_training/reports/";
+
+    @Value("${data.generator.enabled}")
+    private boolean dataGeneratorEnabled;
+
     @PostConstruct
     public void initializeData() throws IOException {
+        if (!dataGeneratorEnabled) {
+            System.out.println("Data generator is disabled. Skipping data initialization.");
+            return;
+        }
+
 
         employeeRepository.deleteAll();
         adminRepository.deleteAll();
